Tokenizer produces only string values
and IPAddress is a configurable type.

+++ Grbl_Esp32/src/Configuration/RuntimeSetting.h
@@ -56,7 +56,7 @@ namespace Configuration {
         RuntimeSetting(const char* runtimeSetting);
 
         void handle(const char* name, int32_t& value, int32_t minValue, int32_t maxValue) override;
-        void handle(const char* name, double& value, double minValue, double maxValue) override;
+        void handle(const char* name, float& value, float minValue, float maxValue) override;
         void handle(const char* name, StringRange& value, int minLength, int maxLength) override;
         void handle(const char* name, Pin& value) override;
         void handle(const char* name, int& value, EnumItem* e) override {}
